% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtm_afg.R
\docType{data}
\name{dtm_afg}
\alias{dtm_afg}
\title{DTM Afghanistan Flow Monitoring Registry}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 124 rows and 6 columns.
}
\source{
DTM Afghanistan.
}
\usage{
dtm_afg
}
\description{
Monthly migrant movements in and out of Afganistan as observed from flow
monitoring points (FMP), by sex and age.
}
\details{
\describe{
\item{t}{month-year of count}
\item{q}{quarter-year of count}
\item{direction}{outflow, inflow}
\item{sex}{sex of migrant}
\item{age}{above 5 years old, below 5 years old}
\item{n}{count of individuals}
}

Received from DTM Afghanistan. Currently available from Jan 2024 to
Jun 2024. Last updated 1 August 2024.
}
\keyword{datasets}
