% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtm_ehoa_dmg.R
\docType{data}
\name{dtm_ehoa_dmg}
\alias{dtm_ehoa_dmg}
\title{DTM East and Horn of Africa Flow Monitoring Registry: Demographics}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 144000 rows and 28 columns.
}
\source{
DTM East and Horn of Africa.
}
\usage{
dtm_ehoa_dmg
}
\description{
Monthly migrant transits as observed from flow monitoring points (FMP) in
East and Horn of Africa countries, by origin, transit, and destination
countries and sex and age. See \code{dtm_ehoa_nat} for the same dataset with
breakdown by nationality.
}
\details{
\describe{
\item{t}{month-year of count}
\item{q}{quarter-year of count}
\item{fmp}{flow monitoring point}
\item{fmp_country}{country of flow monitoring point (ISO-3)}
\item{route}{route as defined by DTM EHOA}
\item{subroute}{subroute as defined by DTM EHOA}
\item{movement}{outgoing, incoming, undefined}
\item{origin}{country of origin (ISO-3)}
\item{previous}{previous country in journey (ISO-3)}
\item{next}{next country in journey (ISO-3)}
\item{final}{intended final destination (ISO-3)}
\item{sex}{sex of migrant}
\item{age}{adults, children, children unaccompanied}
\item{n}{count of individuals}
\item{uac}{count of unaccompanied children}
\item{reason_buy_goods}{a reason for movement is to buy goods for personal consumption (\verb{1=}yes)}
\item{reason_collect_aid}{a reason for movement is to collect aid (\verb{1=}yes)}
\item{reason_conflict}{a reason for movement are conflicts (\verb{1=}yes)}
\item{reason_disaster}{a reason for movement are natural disasters (\verb{1=}yes)}
\item{reason_economic}{a reason for movement is economic (\verb{1=}yes)}
\item{reason_education}{a reason for movement is education (\verb{1=}yes)}
\item{reason_family}{a reason for movement is to visit family (\verb{1=}yes)}
\item{reason_food}{a reason for movement is food insecurity (\verb{1=}yes)}
\item{reason_health}{a reason for movement is health care (\verb{1=}yes)}
\item{reason_living_condition}{a reason for movement is uncomfortable living conditions (\verb{1=}yes)}
\item{reason_seasonal}{a reason for movement is seasonal (\verb{1=}yes)}
\item{reason_services}{a reason for movement is to seek better services (\verb{1=}yes)}
\item{reason_tourism}{a reason for movement is tourism (\verb{1=}yes)}
}

Downloaded from the DTM East and Horn of Africa data server.
Currently available from Jan 2018 to Jul 2024.

The \code{uac} column only has values for rows with \verb{age=}"children". It is the
count of unaccompanied children among total children. Note that \code{uac=NA}
means no information and does not necessarily imply \code{uac=0}.

The \verb{reason_*} columns are not exhaustive. That is, there are rows for
which all columns have value \code{0}, indicating other reasons for movement not
listed.

Last updated 13 January 2025.
}
\keyword{datasets}
