% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtm_fms_afg.R
\docType{data}
\name{dtm_fms_afg}
\alias{dtm_fms_afg}
\title{DTM Flow Monitoring Surveys: Afghanistan}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 158603 rows and 21 columns.
}
\source{
DTM Afghanistan.
}
\usage{
dtm_fms_afg
}
\description{
Responses from DTM Flow Monitoring Surveys conducted in Afghanistan.
}
\details{
\describe{
\item{t}{date of survey}
\item{fmp}{flow monitoring point where survey was conducted}
\item{direction}{outflow, inflow}
\item{nationality}{nationality of migrant (ISO-3)}
\item{sex}{sex of survey respondent}
\item{age}{age of survey respondent}
\item{uam}{unaccompanied minor (\verb{1=}yes)}
\item{origin}{origin of migrant (ISO-3)}
\item{origin_adm1}{province origin of migrant}
\item{origin_adm2}{district origin of migrant}
\item{destination}{destination of migrant (ISO-3)}
\item{destination_adm1}{province destination of migrant}
\item{destination_adm2}{district destination of migrant}
\item{stay_origin}{length of stay in origin country}
\item{stay_destination}{planned length of stay in destination country}
\item{reason_disaster}{a reason for movement is "Natural disaster" (\verb{1=}yes)}
\item{reason_economic}{a reason for movement is "Economic reasons" (\verb{1=}yes)}
\item{reason_education}{a reason for movement is "Education" (\verb{1=}yes)}
\item{reason_family}{a reason for movement is "Family-related visit" (\verb{1=}yes)}
\item{reason_health}{a reason for movement is "Health reasons" (\verb{1=}yes)}
\item{reason_safety}{a reason for movement is "Felt unsafe at the place where I'm coming from" (\verb{1=}yes)}
}

Surveys were conducted in Afghanistan from Jan 2024 to Jul 2024.
Last updated 1 August 2024.

The \verb{reason_*} columns are not exhaustive. That is, there are rows for
which all columns have value \code{0}, indicating other reasons for movement not
listed.
}
\keyword{datasets}
