% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtm_fms_wca.R
\docType{data}
\name{dtm_fms_wca}
\alias{dtm_fms_wca}
\title{DTM Flow Monitoring Surveys: West and Central Africa}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 236378 rows and 26 columns.
}
\source{
DTM West and Central Africa.
}
\usage{
dtm_fms_wca
}
\description{
Responses from DTM Flow Monitoring Surveys conducted in 10 countries in West
and Central Africa.
}
\details{
\describe{
\item{t}{date of survey}
\item{fmp}{flow monitoring point where survey was conducted}
\item{fmp_country}{country of flow monitoring point (ISO-3)}
\item{origin}{origin country (ISO-3)}
\item{previous}{previous country (ISO-3)}
\item{next}{next destination country (ISO-3)}
\item{final}{final destination country (ISO-3)}
\item{route}{migratory route}
\item{nat}{nationality (ISO-3) of respondent}
\item{birth}{country of birth (ISO-3) of respondent}
\item{sex}{sex of respondent}
\item{age}{age of respondent}
\item{reason_*}{=1 if given reason is cited by respondent}
}

Downloaded from the DTM West and Central Africa data server.
Currently available from Jul 2020 to Jun 2024.

Migratory routes are defined by \code{origin} and \code{final} countries. There are
currently three: "Intra-WCA", "WCA to Europe", and "WCA to Others". If
either \code{origin} or \code{final} is unknown (\code{=}"OOO"), then \code{route} is left
blank.

Last updated 10 October 2024.
}
\keyword{datasets}
