% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtm_hnd.R
\docType{data}
\name{dtm_hnd}
\alias{dtm_hnd}
\title{Migrant Arrivals to Honduras}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 42 rows and 3 columns.
}
\source{
Honduras National Institute for Migration; DTM Americas.
}
\usage{
dtm_hnd
}
\description{
Migrant arrivals to Honduras as reported by the government.
}
\details{
\describe{
\item{t}{month-year of count}
\item{q}{quarter-year of count}
\item{n}{count of individuals}
}

Scraped from the Honduras National Institute for Migration's \emph{Flujo
Migratorio Irregular} dashboard.
}
\keyword{datasets}
