% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtm_med.R
\docType{data}
\name{dtm_med}
\alias{dtm_med}
\title{Migrant Arrivals to Europe}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 928 rows and 4 columns.
}
\source{
DTM Europe
}
\usage{
dtm_med
}
\description{
Migrant arrivals to Bulgaria, Cyprus, Greece, Italy, Malta, and Spain.
}
\details{
\describe{
\item{t}{month-year of count}
\item{q}{quarter-year of count}
\item{geo}{reference country (ISO-3)}
\item{n}{count of individuals}
}

Taken from DTM Europe's \emph{Mixed Migration Flows to Europe by Sea and
Land} dashboard
(https://dtm.iom.int/content/europe-migration-arrivals-dashboard). Longest
series is available from Dec 1997 to Dec 2024, but availability varies
across countries. Last updated 16 January 2025.

See \code{dtm_med_nat} for arrivals by country of origin and \code{dtm_med_dmg} for
arrivals by sex, and age.
}
\keyword{datasets}
