% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtm_med_dmg.R
\docType{data}
\name{dtm_med_dmg}
\alias{dtm_med_dmg}
\title{Migration Flows to Europe: Demographics}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 405 rows and 5 columns.
}
\source{
DTM Europe.
}
\usage{
dtm_med_dmg
}
\description{
Migrant arrivals to Italy by demographic.
}
\details{
\describe{
\item{t}{month-year of count}
\item{q}{quarter-year of count}
\item{geo}{arrival country (ISO-3)}
\item{demog}{men, women, children, uac}
\item{n}{count of individuals}
}

Received from DTM Europe. Currently available from Jan 2018 to Sep
2024. Last updated 22 Oct 2024.

Numbers are estimates whose monthly totals are coerced to accord with
\code{dtm_med}, hence the presence of negatives.
}
\keyword{datasets}
