% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtm_med_nat.R
\docType{data}
\name{dtm_med_nat}
\alias{dtm_med_nat}
\title{Migration Flows to Europe: Nationality}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 14609 rows and 5 columns.
}
\source{
DTM Europe; European Border and Coast Guard Agency (Frontex).
}
\usage{
dtm_med_nat
}
\description{
Migrant arrivals to Bulgaria, Cyprus, Greece, Italy, Malta, and Spain by
nationality.
}
\details{
\describe{
\item{t}{month-year of count}
\item{q}{quarter-year of count}
\item{geo}{arrival country (ISO-3)}
\item{nat}{nationality of migrant (ISO-3)}
\item{n}{count of individuals}
}

Taken from DTM Europe's \emph{Mixed Migration Flows to Europe by Sea and
Land} dashboard
(https://dtm.iom.int/content/europe-migration-arrivals-dashboard), except
for arrivals to Spain via the West African Atlantic Route ("ESP-WAAR") and
the Western Mediterranean Route ("ESP-WMR"), which are from Frontex
(https://www.frontex.europa.eu/what-we-do/monitoring-and-risk-analysis/migratory-map).
WAAR covers arrivals by sea to the Canary Islands while WMR covers land and
sea arrivals to Spain excluding the Canary Islands.

Currently available from Jan 2015 to Aug 2024. Last updated 16 January
2025. Total arrivals may not match those in \code{dtm_med}. Where discrepancies
exist, values in \code{dtm_med} take precedence.

See \code{dtm_med_dmg} for arrivals by sex and age.
}
\keyword{datasets}
