% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtm_wca_dmg.R
\docType{data}
\name{dtm_wca_dmg}
\alias{dtm_wca_dmg}
\title{DTM West and Central Africa Flow Monitoring Registry: Demographics}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1877131 rows and 10 columns.
}
\source{
DTM West and Central Africa.
}
\usage{
dtm_wca_dmg
}
\description{
Daily migrant transits as observed from flow monitoring points (FMP) in West
and Central Africa countries, by origin and destination countries and sex and
age.
}
\details{
\describe{
\item{t}{date of count}
\item{route}{migratory route}
\item{fmp}{flow monitoring point}
\item{fmp_country}{country of flow monitoring point (ISO-3)}
\item{from}{departure country (ISO-3)}
\item{to}{destination country (ISO-3)}
\item{sex}{sex of migrant}
\item{age}{18+ or under 18}
\item{n}{count of individuals}
\item{uac}{count of unaccompanied children among under 18}
}

Downloaded from the DTM West and Central Africa data server.
Currently available from Jan 2018 to Jul 2024.

Migratory routes are defined as a pair of adjacent WCA countries. Transits
are both ways.

The \code{uac} column only has values for rows with \verb{age=}"under 18". It is the
count of unaccompanied children among total children. Note that \code{uac=NA}
means no information and does not necessarily imply \code{uac=0}.

Last updated 1 October 2024.
}
\keyword{datasets}
