% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsi.R
\docType{data}
\name{fsi}
\alias{fsi}
\title{Fragile States Index}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 38396 rows and 7 columns.
}
\source{
The Fund for Peace
}
\usage{
fsi
}
\description{
The Fragile States Index (FSI) is produced by The Fund for Peace (FFP) and
provides a measure of state pressure and the sources of pressure
}
\details{
\describe{
\item{geo}{country, area or territory (ISO-3)}
\item{t}{year}
\item{rank}{rank of country for given year}
\item{total}{overall index}
\item{code}{code of sub-indicator}
\item{indicator}{sub-indicator name}
\item{v}{value}
}

See https://fragilestatesindex.org/indicators for a description of each
sub-indicator.

Last updated 15 May 2024.
}
\keyword{datasets}
