% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idmc.R
\docType{data}
\name{idmc_flows}
\alias{idmc_flows}
\title{Global Internal Displacement Database: Flows}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4071 rows and 4 columns.
}
\source{
Internal Displacement Monitoring Centre (private correspondence).
}
\usage{
idmc_flows
}
\description{
Flow count of internally displaced persons by cause of displacement.
}
\details{
\describe{
\item{geo}{country, area, or territory (ISO-3)}
\item{t}{year}
\item{cause}{cause of displacement}
\item{n}{count of individuals}
}

Received via correspondence with IDMC.

Numbers represent the number of internal displacements, or internal
displacement population flows, reported from January 1st to December 31st of
a reporting year. This figure may include individuals who are displaced
multiple times during the year by different events.

Conflict displacements are situations where people are forced to leave their
homes or places of habitual residence as a result or in order to avoid the
impact of armed conflict, communal violence and criminal violence.

Disaster displacements are situations where people are forced to leave their
homes or places of habitual residence as a result, or in anticipation of the
negative impact of natural hazards. Disasters include:
\itemize{
\item Drought
\item Earthquake
\item Erosion
\item Extreme temperature
\item Flood
\item Mass movement
\item Storm
\item Volcanic activity
\item Wave action
\item Wildfire
}

IDMC separately identifies the Abyei Area (\verb{geo=}"AB9"), a disputed area on
the border between Sudan and South Sudan.
}
\keyword{datasets}
