% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irreg_australia.R
\docType{data}
\name{irreg_australia}
\alias{irreg_australia}
\title{Australia Boat Arrivals}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 50 rows and 3 columns.
}
\source{
Refugee Council of Australia.
}
\usage{
irreg_australia
}
\description{
Statistics on refugee boats arrivals to Australia.
}
\details{
\describe{
\item{t}{year}
\item{boats}{count of boats}
\item{n}{count of individuals}
}

Downloaded from
https://www.refugeecouncil.org.au/asylum-boats-statistics. Period covered
is 1976 to 2024, and Jan-Apr 2025.

The terms "boat arrivals" here refers to people on board the vessels that
could not be safely returned and therefore landed in Australia as well as
those who were returned to countries of departure or origin.

The figures for 2022 onward includes all passengers and crew, as no
breakdown was provided. Note that for the latest year, figures of less than
5 are rendered exact, until confirmed by later data.

Last updated 19 Jun 2025.
}
\keyword{datasets}
