% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irreg_mexico.R
\docType{data}
\name{irreg_mexico_dmg}
\alias{irreg_mexico_dmg}
\title{People in an Irregular Migratory Situation in Mexico: Demographics}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 720 rows and 4 columns.
}
\source{
México Unidad de Política Migratoria [Migration Policy Unit ].
}
\usage{
irreg_mexico_dmg
}
\description{
Count of foreigners in Mexico for whom a \emph{Procedimiento Administrativo
Migratorio} (PAM) was initiated before the \emph{Instituto Nacional de Migración}
(INM) for not accrediting their immigration status and those who were
channeled by this same authority to the shelters of the \emph{Sistema Nacional para
el Desarrollo Integral de la Familia} (DIF) network, with PAM initiated, by
sex and age. See \code{irreg_mexico_nat} for the same dataset with breakdown by country
of origin.
}
\details{
\describe{
\item{t}{month-year of count}
\item{sex}{male, female}
\item{age}{age range}
\item{n}{count of individuals}
}

For each year of the \emph{Boletines Estadísticos}, table 3.1.3 was
downloaded from
https://portales.segob.gob.mx/es/PoliticaMigratoria/Boletines_Estadisticos.
Period covered is currently Jan 2018 to Jun 2025.

Possible values for \code{age} are "under 12, accompanied", "under 12,
unaccompanied", "12-17" and "18+".

Last updated 18 Aug 2025.
}
\keyword{datasets}
