% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/undesa_stocks.R
\docType{data}
\name{undesa_stocks}
\alias{undesa_stocks}
\title{International Migrant Stocks}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 153296 rows and 9 columns.
}
\source{
United Nations Department of Economic and Social Affairs.
}
\usage{
undesa_stocks
}
\description{
Count of migrant population in countries, areas, and territories by place of
origin and sex, as estimated by UN DESA.
}
\details{
\describe{
\item{t}{year}
\item{geo}{country, area, or territory (ISO-3)}
\item{from}{origin of migrants (ISO-3)}
\item{sex}{sex of migrants}
\item{n}{count}
\item{type_b}{count is based on foreign-born population}
\item{type_c}{count is based on foreign citizens}
\item{type_r}{count includes refugees}
\item{type_i}{no data exists so count has been imputed}
}

UN DESA estimates international migrant stocks from national
statistics, usually population censuses, as well as nationally
representative surveys. Data is available at 5-year intervals from 1990 to
2020, and then for 2024. Counts are as of July 1st of the given year.
Estimates are as of 27 January 2024.

Note that \code{from}–\code{sex} combinations that are not present for a given
country means that insufficient data exist to estimate their presence, not
that the particular \code{from}–\code{sex} combination is absent. Only cases with
\code{n=0} mean a definite absence.

For statistical purposes, UN DESA treats the following as separate
entities:
\itemize{
\item The Regions of Metropolitan France; French Guiana; French
Polynesia; Guadeloupe; Martinique; Mayotte; New Caledonia; Réunion;
Saint Pierre and Miquelon; Saint Barthélemy; Saint Martin (French
part); Wallis and Futuna Islands.
\item New Zealand; Cook Islands; Niue; Tokelau.
\item Denmark; Faroe Islands; Greenland.
\item The Netherlands; Aruba; Bonaire, Sint Eustatius and Saba; Curaçao;
Sint Maarten (Dutch part).
\item The United Kingdom; Anguilla; Bermuda; British Virgin Islands;
Cayman Islands; Channel Islands; Falkland Islands (Malvinas);
Gibraltar; Isle of Man; Montserrat; Saint Helena; Turks and Caicos
Islands.
\item The 50 states of the United States of America and the District of
Columbia; Puerto Rico; Guam; Northern Mariana Islands; United States
Virgin Islands.
\item China; Hong Kong Special Administrative Region, China; Macao
Special Administrative Region, China; Taiwan Province of China.
}

UN DESA also notes the following for particular territories:
\itemize{
\item \strong{Bonaire, Sint Eustatius and Saba}: Estimates for 1990 to 2000
refer to the former Netherlands Antilles. Estimates for 2005 refer to
the former Netherlands Antilles without Sint Maarten (Dutch part).
\item \strong{Occupied Palestinian Territory}: Includes East Jerusalem.
Refugees are not part of the foreign-born migrant stock.
\item \strong{Sudan}: Estimates for 1990 to 2005 refer to Sudan and South
Sudan.
}
}
\keyword{datasets}
