% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/undesa_wpp.R
\docType{data}
\name{undesa_wpp}
\alias{undesa_wpp}
\title{World Population Prospects}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3555000 rows and 5 columns.
}
\source{
United Nations Department of Economic and Social Affairs.
}
\usage{
undesa_wpp
}
\description{
Population count (in thousands) by sex and age from the UN World Population
Prospects 2024 update.
}
\details{
Downloaded from
https://population.un.org/wpp/downloads?folder=Standard\%20Projections&group=CSV\%20format
on 21 January 2025.

For statistical purposes, UN DESA treats the following as separate
entities:
\itemize{
\item The Regions of Metropolitan France; French Guiana; French
Polynesia; Guadeloupe; Martinique; Mayotte; New Caledonia; Réunion;
Saint Pierre and Miquelon; Saint Barthélemy; Saint Martin (French
part); Wallis and Futuna Islands.
\item New Zealand; Cook Islands; Niue; Tokelau.
\item Denmark; Faroe Islands; Greenland.
\item The Netherlands; Aruba; Bonaire, Sint Eustatius and Saba; Curaçao;
Sint Maarten (Dutch part).
\item The United Kingdom; Anguilla; Bermuda; British Virgin Islands;
Cayman Islands; Channel Islands; Falkland Islands (Malvinas);
Gibraltar; Isle of Man; Montserrat; Saint Helena; Turks and Caicos
Islands.
\item The 50 states of the United States of America and the District of
Columbia; Puerto Rico; Guam; Northern Mariana Islands; United States
Virgin Islands.
\item China; Hong Kong Special Administrative Region, China; Macao
Special Administrative Region, China; Taiwan Province of China.
}

UN DESA also notes the following for particular territories:
\itemize{
\item \strong{Bonaire, Sint Eustatius and Saba}: Estimates for 1990 to 2000
refer to the former Netherlands Antilles. Estimates for 2005 refer to
the former Netherlands Antilles without Sint Maarten (Dutch part).
\item \strong{Occupied Palestinian Territory}: Includes East Jerusalem.
Refugees are not part of the foreign-born migrant stock.
\item \strong{Sudan}: Estimates for 1990 to 2005 refer to Sudan and South
Sudan.
}

Data last updated 11 Jul 2024.
}
\keyword{datasets}
