% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eu_asyapp.R
\docType{data}
\name{eu_asyapp}
\alias{eu_asyapp}
\title{EU Asylum Applications}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 43896 rows and 5 columns.
}
\source{
Eurostat.
}
\usage{
eu_asyapp
}
\description{
First time asylum application in the EU, by country of application and applicant's
nationality and sex.
}
\details{
\describe{
\item{geo}{Eurostat country (ISO-3)}
\item{t}{sea or land}
\item{nat}{nationality of applicant (ISO-3)}
\item{sex}{male, female, unknown}
\item{n}{count}
}

Available from 2011 to 2024. Downloaded through the Eurostat API
with the dataset key "migr_asyappctza".

Countries of applications are:
\itemize{
\item Austria
\item Belgium
\item Bulgaria
\item Croatia
\item Cyprus
\item Czechia
\item Denmark
\item Germany
\item Estonia
\item Finland
\item France
\item Greece
\item Hungary
\item Ireland
\item Iceland
\item Italy
\item Liechtenstein
\item Lithuania
\item Luxembourg
\item Latvia
\item Malta
\item Montenegro
\item Kingdom of the Netherlands
\item Norway
\item Poland
\item Portugal
\item Romania
\item Slovakia
\item Slovenia
\item Spain
\item Sweden
\item Switzerland
\item United Kingdom
\item EU27 (aggregation of all EU countries)
}

Last updated 26 May 2025.
}
\keyword{datasets}
