% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurostat.R
\docType{data}
\name{eu_asyapp_m}
\alias{eu_asyapp_m}
\title{EU Asylum Applications (Monthly)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2179 rows and 3 columns.
}
\source{
Eurostat.
}
\usage{
eu_asyapp_m
}
\description{
First time asylum application in the EU, by applicant's nationality.
}
\details{
\describe{
\item{t}{year-month}
\item{nat}{nationality of applicant (ISO-3)}
\item{n}{count}
}

Available from Jan 2024 to May 2025. Downloaded through the Eurostat
API with the dataset key "migr_asyappctzm".

Last updated 18 Aug 2025.
}
\keyword{datasets}
