% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurostat.R
\docType{data}
\name{eu_eiord}
\alias{eu_eiord}
\title{EU Third-Country Nationals Ordered to Leave}
\format{
A dataset with variables:
\describe{
\item{t}{year-quarter}
\item{geo}{Reporting country (ISO-3)}
\item{nat}{nationality of applicant (ISO-3)}
\item{n}{count}
}
}
\source{
Eurostat <migr_eiord1> (downloaded 21 Oct 2025).
}
\usage{
eu_eiord
}
\description{
Quarterly count of third-country nationals ordered to leave in European
countries, by nationality. Available from 2021 Q1 to 2025 Q3.
}
\details{
The column \code{geo} includes "EU27", which the aggregate count for the
27 European Countries.
}
\keyword{datasets}
