% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurostat.R
\docType{data}
\name{eu_eiord}
\alias{eu_eiord}
\title{EU Third-Country Nationals Ordered to Leave}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 26363 rows and 4 columns.
}
\source{
Eurostat.
}
\usage{
eu_eiord
}
\description{
Quarterly count of third-country nationals ordered to leave in European
countries, by nationality.
}
\details{
\describe{
\item{t}{year-quarter}
\item{geo}{Reporting country (ISO-3)}
\item{nat}{nationality of applicant (ISO-3)}
\item{n}{count}
}

Available from 2021 Q1 to 2025 Q1. Last updated 10 Jul 2025.

The column \code{geo} includes "EU27", which the aggregate count for the 27
European Countries.
}
\keyword{datasets}
