% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurostat.R
\docType{data}
\name{eu_eirtn}
\alias{eu_eirtn}
\title{EU Third-Country Nationals Returned Following an Order to Leave}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 22338 rows and 5 columns.
}
\source{
Eurostat.
}
\usage{
eu_eirtn
}
\description{
Quarterly count of third-country nationals returned following an order to
leave in European countries, by nationality and type of return.
}
\details{
\describe{
\item{t}{year-quarter}
\item{geo}{Reporting country (ISO-3)}
\item{nat}{nationality of applicant (ISO-3)}
\item{dest}{type of return}
\item{n}{count}
}

Available from 2021 Q1 to 2025 Q1. Last updated 10 Jul 2025.

The column \code{geo} includes "EU27", which the aggregate count for the 27
European Countries.

The column \code{dest} takes the following values:
\itemize{
\item "EU27_2020_EFTA_FOR": EU27 or EFTA countries other than reporting country
\item "THRD_CTZ": country of returnee's nationality
\item "THRD_OTH": other third countries
\item "UNK": unknown
}
}
\keyword{datasets}
