% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iom_mmp.R
\docType{data}
\name{iom_mmp}
\alias{iom_mmp}
\title{Missing Migrants Project}
\format{
A dataset with variables:
\describe{
\item{t}{date of incident}
\item{m}{month-year of incident}
\item{type}{incident, split, cumulative}
\item{geo}{country of incident (ISO-3)}
\item{dead}{number of dead or missing}
\item{cause}{cause of death}
\item{survive}{number of survivors (if available)}
\item{nat}{country/ies of origin of dead and missing}
\item{location}{precise location of incident}
\item{lon}{longitude of incident}
\item{lat}{latitude of incident}
}
}
\source{
IOM Missing Migrants Project
\url{https://missingmigrants.iom.int/downloads} (downloaded 25 Nov 2025).
}
\usage{
iom_mmp
}
\description{
Deaths and disappearances of people in the process of migration towards an
international destination. Covers Jan 2014 to Nov 2025.
}
\keyword{datasets}
