% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irreg_darien.R
\docType{data}
\name{irreg_darien_dmg}
\alias{irreg_darien_dmg}
\title{Irregular Transits Through Darién: Demographics}
\format{
A dataset with variables:
\describe{
\item{t}{month-year of count}
\item{demog}{sex and age category}
\item{n}{count of individuals}
}
}
\source{
Panama Servicio Nacional de Migración [National Migration Service]
\url{https://www.migracion.gob.pa/estadisticas/} (downloaded 28 Nov 2025).
}
\usage{
irreg_darien_dmg
}
\description{
Monthly count of irregular transits through the Darién region of Panama, by
sex and age. Available from Jan 2021 to Sep 2025. See \code{irreg_darien_nat} for
the same dataset with breakdown by nationality.
}
\details{
From 2021 to 2022, demographic categories are only "men" and
"women". From 2023 onwards, categories are "men", "women", "boys" and
"girls".
}
\keyword{datasets}
