% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irreg_darien.R
\docType{data}
\name{irreg_darien_nat}
\alias{irreg_darien_nat}
\title{Irregular Transits Through Darién: Nationality}
\format{
A dataset with variables:
\describe{
\item{t}{month-year of count}
\item{nat}{nationality (ISO-3)}
\item{n}{count of individuals}
}
}
\source{
Panama Servicio Nacional de Migración [National Migration Service]
\url{https://www.migracion.gob.pa/estadisticas/} (downloaded 28 Nov 2025).
}
\usage{
irreg_darien_nat
}
\description{
Monthly count of irregular transits through the Darién region of Panama, by
nationality. Available from 2010-2019 (annual) and Jan 2021 to Sep 2025. See
\code{irreg_darien_dmg} for the same dataset with breakdown by sex and age.
}
\keyword{datasets}
