% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irreg_darien_dmg.R, R/irreg_darien_nat.R
\docType{data}
\name{irreg_darien_nat}
\alias{irreg_darien_nat}
\title{Irregular Transits Through Darien: Demographics}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2268 rows and 4 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2268 rows and 4 columns.
}
\source{
Panama Servicio Nacional de Migración [National Migration Service].

Panama Servicio Nacional de Migración [National Migration Service].
}
\usage{
irreg_darien_nat

irreg_darien_nat
}
\description{
Monthly count of irregular transits through the Darien region of Panama, by
sex and age See \code{irreg_darien_nat} for the same dataset with breakdown by
nationality.

Monthly count of irregular transits through the Darien region of Panama, by
nationality. See \code{irreg_darien_dmg} for the same dataset with breakdown by
sex and age.
}
\details{
\describe{
\item{t}{month-year of count}
\item{q}{quarter-year of count}
\item{demog}{sex and age category}
\item{n}{count of individuals}
}

Downloaded from https://www.migracion.gob.pa/estadisticas. Period
covered is currently Jan 2021 to Dec 2024. From 2021 to 2022, demographic
categories are only "men" and "women". From 2023 onwards, categories are
"men", "women", "boys" and "girls".

Data as of 31 December 2024. Last updated 13 February 2025.

\describe{
\item{t}{month-year of count}
\item{q}{quarter-year of count}
\item{nat}{nationality (ISO-3)}
\item{n}{count of individuals}
}

Downloaded from https://www.migracion.gob.pa/estadisticas. Period
covered is currently 2010-2019 (annual) and Jan 2020 to Dec 2024.

Data as of 31 December 2024. Last updated 12 February 2025.
}
\keyword{datasets}
