% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irreg_darien_nat.R
\docType{data}
\name{irreg_darien_nat}
\alias{irreg_darien_nat}
\title{Irregular Transits Through Darién: Nationality}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2338 rows and 4 columns.
}
\source{
Panama Servicio Nacional de Migración [National Migration Service].
}
\usage{
irreg_darien_nat
}
\description{
Monthly count of irregular transits through the Darién region of Panama, by
nationality. See \code{irreg_darien_dmg} for the same dataset with breakdown by
sex and age.
}
\details{
\describe{
\item{t}{month-year of count}
\item{q}{quarter-year of count}
\item{nat}{nationality (ISO-3)}
\item{n}{count of individuals}
}

Downloaded from https://www.migracion.gob.pa/estadisticas. Period
covered is currently 2010-2019 (annual) and Jan 2020 to Feb 2025.

Data as of 28 Feb 2025.
}
\keyword{datasets}
