% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irreg_guatemala.R
\docType{data}
\name{irreg_guatemala}
\alias{irreg_guatemala}
\title{Guatemalans Returned from Mexico and United States}
\format{
A dataset with variables:
\describe{
\item{t}{month-year of count}
\item{sex}{male, female}
\item{age}{adult, accompanied child, unaccompanied child}
\item{origin}{MEX, USA}
\item{n}{count of individuals}
}
}
\source{
Instituto Guatemalteco de Migración [Guatemalan Institute of
Migration] \url{https://igm.gob.gt/informes-estadisticos} (downloaded 30 Nov
2025).
}
\usage{
irreg_guatemala
}
\description{
Monthly count of Guatemalans returned from Mexico and the United States by
sex and age. Covers Jan 2020 to Oct 2025.
}
\details{
Children are defined as those under 18 years of age.
}
\keyword{datasets}
