% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irreg_guatemala.R
\docType{data}
\name{irreg_guatemala}
\alias{irreg_guatemala}
\title{Guatemalans Returned from Mexico and United States}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 736 rows and 5 columns.
}
\source{
Instituto Guatemalteco de Migración [Guatemalan Institute of
Migration].
}
\usage{
irreg_guatemala
}
\description{
Monthly count of Guatemalans returned from Mexico and the United States by
sex and age.
}
\details{
\describe{
\item{t}{month-year of count}
\item{sex}{male, female}
\item{age}{adult, accompanied child, unaccompanied child}
\item{origin}{MEX, USA}
\item{n}{count of individuals}
}

Children are defined as those under 18 years of age.

Downloaded from https://igm.gob.gt/informes-estadisticos. Period covered is
from 2020 Jan to 2025 Apr. Last updated 18 Jun 2025.
}
\keyword{datasets}
