% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irreg_honduras.R
\docType{data}
\name{irreg_honduras}
\alias{irreg_honduras}
\title{Irregular Migrant Arrivals to Honduras}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 51 rows and 3 columns.
}
\source{
Honduras National Institute for Migration; DTM Americas.
}
\usage{
irreg_honduras
}
\description{
Irregular migrant arrivals to Honduras as reported by the government.
}
\details{
\describe{
\item{t}{month-year of count}
\item{q}{quarter-year of count}
\item{n}{count of individuals}
}

Scraped from the Honduras National Institute for Migration's \emph{Flujo
Migratorio Irregular} dashboard
(https://app.powerbi.com/view?r=eyJrIjoiODNhMGY4MGUtM2NkMy00YTIzLTlmZTAtNmQxZmJiZDhiZjc1IiwidCI6ImUxMWQ5NjIwLTRkM2UtNDEwYi05MTUyLWZkMWNmNmNmNzI5YSJ9).

Last updated 10 Mar 2025.
}
\keyword{datasets}
