% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irreg_honduras.R
\docType{data}
\name{irreg_honduras_nat}
\alias{irreg_honduras_nat}
\title{Irregular Migrant Arrivals to Honduras: Nationalities}
\format{
A dataset with variables:
\describe{
\item{t}{year-month of count}
\item{nat}{nationality}
\item{n}{count of individuals}
}
}
\source{
Honduras National Institute for Migration
\url{https://app.powerbi.com/view?r=eyJrIjoiODNhMGY4MGUtM2NkMy00YTIzLTlmZTAtNmQxZmJiZDhiZjc1IiwidCI6ImUxMWQ5NjIwLTRkM2UtNDEwYi05MTUyLWZkMWNmNmNmNzI5YSJ9}
(accessed 30 Sep 2025); DTM Americas.
}
\usage{
irreg_honduras_nat
}
\description{
Irregular migrant arrivals to Honduras as reported by the government, by
nationality. Covers Jan 2018 to Aug 2025.
}
\keyword{datasets}
