% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irreg_mexico.R
\docType{data}
\name{irreg_mexico_dmg}
\alias{irreg_mexico_dmg}
\title{Irregular Migrants Processed in Mexico: Demographics}
\format{
A dataset with variables:
\describe{
\item{t}{month-year of count}
\item{sex}{male, female}
\item{age}{age range}
\item{n}{count of individuals}
}
}
\source{
México Unidad de Política Migratoria [Migration Policy Unit ];
table 3.1.3 at
\url{https://portales.segob.gob.mx/es/PoliticaMigratoria/Boletines_Estadisticos}
(updated 3 Sep 2025).
}
\usage{
irreg_mexico_dmg
}
\description{
Count of foreigners in Mexico for whom a \emph{Procedimiento Administrativo
Migratorio} (PAM) was initiated before the \emph{Instituto Nacional de Migración}
(INM) for not accrediting their immigration status and those who were
channeled by this same authority to the shelters of the \emph{Sistema Nacional para
el Desarrollo Integral de la Familia} (DIF) network, with PAM initiated, by
sex and age. See \code{irreg_mexico_nat} for the same dataset with breakdown by country
of origin. Covers Jan 2018 to Jul 2025.
}
\details{
Possible values for \code{age} are "under 12, accompanied", "under 12,
unaccompanied", "12-17" and "18+".
}
\keyword{datasets}
