% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irreg_mexico.R
\docType{data}
\name{irreg_mexico_exp}
\alias{irreg_mexico_exp}
\title{Expulsions from Mexico}
\format{
A dataset with variables:
\describe{
\item{t}{month-year of count}
\item{nat}{nationality (ISO-3)}
\item{type}{deportation, assisted return, assisted return of children}
\item{n}{count of individuals}
}
}
\source{
México Unidad de Política Migratoria [Migration Policy Unit ];
table 3.2.4 at
\url{https://portales.segob.gob.mx/es/PoliticaMigratoria/Boletines_Estadisticos}
(updated 28 Oct 2025).
}
\usage{
irreg_mexico_exp
}
\description{
Count of foreigners expelled from Mexico either through deportation or
assisted return, by nationality. Covers Jan 2018 to Sep 2025.
}
\keyword{datasets}
