% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irreg_mexico.R
\docType{data}
\name{irreg_mexico_nat}
\alias{irreg_mexico_nat}
\title{Irregular Migrants Processed in Mexico: Nationality}
\format{
A dataset with variables:
\describe{
\item{t}{month-year of count}
\item{nat}{nationality (ISO-3)}
\item{n}{count of individuals}
}
}
\source{
México Unidad de Política Migratoria [Migration Policy Unit ];
table 3.1.1 at
\url{https://portales.segob.gob.mx/es/PoliticaMigratoria/Boletines_Estadisticos}
(updated 28 Oct 2025).
}
\usage{
irreg_mexico_nat
}
\description{
Count of foreigners in Mexico for whom a \emph{Procedimiento Administrativo
Migratorio} (PAM) was initiated before the \emph{Instituto Nacional de Migración}
(INM) for not accrediting their immigration status and those who were
channeled by this same authority to the shelters of the \emph{Sistema Nacional
para el Desarrollo Integral de la Familia} (DIF) network, with PAM initiated,
by nationality. See \code{irreg_mexico_dmg} for the same dataset with breakdown by
sex and age. Covers Jan 2018 to Sep 2025.
}
\keyword{datasets}
