% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irreg_mexico_nat.R
\docType{data}
\name{irreg_mexico_nat}
\alias{irreg_mexico_nat}
\title{Mexico Immigrants in an Irregular Situation: Nationality}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5496 rows and 4 columns.
}
\source{
México Unidad de Política Migratoria [Migration Policy Unit ].
}
\usage{
irreg_mexico_nat
}
\description{
Count of foreigners in Mexico for whom a \emph{Procedimiento Administrativo
Migratorio} (PAM) was initiated before the \emph{Instituto Nacional de Migración}
(INM) for not accrediting their immigration status and those who were
channeled by this same authority to the shelters of the \emph{Sistema Nacional
para el Desarrollo Integral de la Familia} (DIF) network, with PAM initiated,
by nationality. See \code{irreg_mexico_dmg} for the same dataset with breakdown
by sex and age.
}
\details{
\describe{
\item{t}{month-year of count}
\item{q}{quarter-year of count}
\item{nat}{nationality (ISO-3)}
\item{n}{count of individuals}
}

For each year of the \emph{Boletines Estadísticos}, table 3.1.1 was
downloaded from
https://portales.segob.gob.mx/es/PoliticaMigratoria/Boletines_Estadisticos.
Period covered is currently Jan 2018 to Aug 2024.

Last updated 17 October 2024.
}
\keyword{datasets}
