% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irreg_usa.R
\docType{data}
\name{irreg_usa}
\alias{irreg_usa}
\title{U.S. Southwest Land Border Encounters}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 37499 rows and 9 columns.
}
\source{
U.S. Customs and Border Protection
\url{https://www.cbp.gov/document/stats/nationwide-encounters} (as of 5
September 2025).
}
\usage{
irreg_usa
}
\description{
Monthly encounters by border authorities of noncitizens irregularly crossing
the United States southwest border, by demographics and nationality. Covers
Oct 2017 to Aug 2025.
}
\details{
\describe{
\item{t}{month-year of encounter}
\item{t_fy}{month-year of encounter in fiscal year}
\item{component}{division under the U.S. Customs and Border Protection involved in the encounter, either the U.S. Border Patrol (UBP) or the Office of Field Operations (OFO)}
\item{aor}{area of responsibility of \code{component}}
\item{authority}{title of authority the noncitizen was processed}
\item{type}{apprehension, inadmissible, or expulsion}
\item{demog}{demographics}
\item{nat}{nationality in ISO-3, available from Oct 2019 onwards}
\item{n}{count of individuals}
}

Data is originally reported by fiscal year, which begins in October.
The column \code{t} is the calendar year equivalents.

\code{area} and \code{nat} are available from Oct 2019 onward.

\code{authority} is either Title 8 or Title 42 of the U.S. Code. Encounters of
noncitizens are generally processed under Title 8, encompassing laws on
"Aliens and Nationality". A public health order during the COVID-19
pandemic allowed for the immediate expulsion of noncitizens under Title 42.
Except for a temporary suspension from 15 November to 19 December 2022,
this policy was in force from 21 March 2020 to 11 May 2023.

Possible values for \code{demog} are "Single Adults", "FMUA" (individuals in a
family unit), "Accompanied Minors" and "UC / Single Minors". FMUA
represents the number of encountered individuals within the family unit,
defined as at least one noncitizen minor child and at least one noncitizen
legal guardian. Accompanied minors represent inadmissible children
accompanied by a legal guardian who is either a U.S. citizen, lawful
permanent resident, or admissible noncitizen. Accompanied minors is a
demographic used only by OFO.
}
\keyword{datasets}
