% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unhcr.R
\docType{data}
\name{unhcr}
\alias{unhcr}
\title{Forcibly Displaced Population}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 191857 rows and 5 columns.
}
\source{
United Nations High Commissioner for Refugees.
}
\usage{
unhcr
}
\description{
Count of forcibly displaced populations at yearend, as reported by UNHCR.
}
\details{
\describe{
\item{geo}{country, area, or territory of asylum (ISO-3)}
\item{t}{year}
\item{from}{origin of forcibly displaced persons (ISO-3)}
\item{group}{refugee, asylum, refugee return, idp, idp return, stateless, other}
\item{n}{count}
}

Downloaded from
https://www.unhcr.org/refugee-statistics/download/?url=tzZ1du on 4
September 2024. Data was last updated 8 October 2024.

This dataset counts populations who are in need of international
protection, classified into four groups: refugees ("refugee"),
asylum-seekers ("asylum"), internally displaced persons ("idp"), and others
("other").

The group "refugee return" are persons returned from abroad after seeking
international protection. These comprise repatriating refugees,
repatriating asylum-seekers whose asylum application has been decided
negative, and other groups who sought or received forms of international
protection and returned to their former country of habitual residence. The
corresponding group for IDPs are "idp return".

In compiling these statistics, UNHCR follows the \emph{International
Recommendations on Refugee Statistics} (2018), an internationally accepted
framework for statistics on refugee and refugee-related populations.

Only conflict-generated IDPs are compiled by UNHCR. See \code{idmc_stocks} for
displacements generated by both conflict and natural disasters.

For "stateless", UNHCR uses the working definition of a person who lacks
proof of possession of any nationality and who at the same time has or is
perceived as having links to a State other than the one he/she is living
in.

Notes on country definitions.

\itemize{
\item Serbia and Kosovo are reported as one entity.
\item China, Hong Kong Special Administrative Region, and Macao
Special Administrative Region are reported as separate entities.
}
}
\keyword{datasets}
